add hook for ordino logo
