+++ src/components/OrdinoLogo.tsx
@@ -1,28 +1,8 @@
 import * as React from 'react';
-import { useAsync, PluginRegistry } from 'tdp_core';
-import { useMemo } from 'react';
-import { EP_ORDINO_LOGO } from '../base';
+import { useOrdinoLogo } from '../hooks/useOrdinoLogo';
 
 export function OrdinoLogo() {
-  const loadOrdinoLogo = useMemo(
-    () => async () => {
-      const defaultSize = { width: 30, height: 30 };
-
-      const plugins = PluginRegistry.getInstance().listPlugins(EP_ORDINO_LOGO);
-      const plugin = plugins?.[0]; // use first registerd plugin; the order depends on import order in the phovea_registry.js of the workspace
-      const module = await (await plugin.load()).factory();
-
-      return {
-        icon: module.default,
-        text: plugin.text,
-        width: plugin.width || defaultSize.width,
-        height: plugin.height || defaultSize.height,
-      };
-    },
-    [],
-  );
-
-  const { status, value } = useAsync(loadOrdinoLogo, []);
+  const { status, value } = useOrdinoLogo();
   return (
     status === 'success' && (
       <div className="ordino-logo">
